changed some h to b

+++ precision_test.cu
@@ -24,7 +24,7 @@ __global__ void axpy(int n, half a, half* x, half* y){
 
 __global__ void axpy(int n, nv_bfloat16 a, nv_bfloat16* x, nv_bfloat16* y){
 	int i = blockIdx.x*blockDim.x + threadIdx.x;
-	half b = __hmul(a,x[i]);
+	nv_bfloat16 b = __hmul(a,x[i]);
 	if (i < n){y[i] = __hadd(b,y[i]);}
 }
